// types/companies.ts
export type Company = {
  id: string;
  name: string;
  logo_url: string | null;
  website: string | null;
  location: string | null;
  industry: string | null;
  description: string | null;
  founded_year: number | null;
  company_size: string | null;
  cover_image:string | null;
  employees:string | null ;
  headquarters: string | null;
  careers_url: string | null;
  linkedin_url: string |null;
  twitter_url: string |null;
  rating: number | null;
  open_jobs_count: number | null;
};